add ipc and more refactoring (#3)
* add logrecover for goroutines and kill process group for monitor

* better logging

* add idle period option

* little cleanup and remove global in desktop

* 1st pass at ipc

* by def only count printable keys for typing pattern

* allow stop/start base run handler for remote calls

* pass ctx/log to run instead of preset

* better out/err timestamps

* use private runner for idle patterns

* more logrecovers for safety

* req >1 keypresses to stop idle

+++ cmd/pattern.go
@@ -1,12 +1,15 @@
 package cmd
 
 import (
+	"context"
 	"fmt"
 	"os"
+	"path/filepath"
 	"strconv"
 	"syscall"
 
 	"github.com/BitPonyLLC/huekeys/buildinfo"
+	"github.com/BitPonyLLC/huekeys/pkg/ipc"
 	"github.com/BitPonyLLC/huekeys/pkg/patterns"
 
 	"github.com/rs/zerolog/log"
@@ -22,12 +25,14 @@ var runCmd = &cobra.Command{
 func init() {
 	rootCmd.AddCommand(runCmd)
 
-	runCmd.PersistentFlags().String("pidpath", os.TempDir()+buildinfo.Name+".pid", "pathname of the pidfile")
+	defaultPidPath := filepath.Join(os.TempDir(), buildinfo.Name+".pid")
+	runCmd.PersistentFlags().String("pidpath", defaultPidPath, "pathname of the pidfile")
 	viper.BindPFlag("pidpath", runCmd.PersistentFlags().Lookup("pidpath"))
 
 	runCmd.PersistentFlags().Int("nice", 10, "the priority level of the process")
 	viper.BindPFlag("nice", runCmd.PersistentFlags().Lookup("nice"))
 
+	addPatternCmd("wait for remote commands", patterns.NewWaitPattern())
 	addPatternCmd("pulse the keyboard brightness up and down", patterns.NewPulsePattern())
 	addPatternCmd("loop through all the colors of the rainbow", patterns.NewRainbowPattern())
 	addPatternCmd("constantly change the color to a random selection", patterns.NewRandomPattern())
@@ -36,12 +41,22 @@ func init() {
 
 	typingPattern := patterns.NewTypingPattern()
 	typingPatternCmd := addPatternCmd("change the color according to typing speed (cold to hot)", typingPattern)
+
 	typingPatternCmd.Flags().String("input-event-id", typingPattern.InputEventID, "input event ID to monitor")
 	viper.BindPFlag("typing.input-event-id", typingPatternCmd.Flags().Lookup("input-event-id"))
-	typingPatternCmd.Flags().StringP("idle", "i", "", "name of pattern to run while keyboard is idle for more than 30 seconds")
+
+	typingPatternCmd.Flags().Bool("all-keys", typingPattern.CountAllKeys, "count any key pressed instead of only those that are considered \"printable\"")
+	viper.BindPFlag("typing.all-keys", typingPatternCmd.Flags().Lookup("all-keys"))
+
+	typingPatternCmd.Flags().StringP("idle", "i", "", "name of pattern to run while keyboard is idle for more than the idle period")
 	viper.BindPFlag("typing.idle", typingPatternCmd.Flags().Lookup("idle"))
+
+	typingPatternCmd.Flags().DurationP("idle-period", "p", typingPattern.IdlePeriod, "amount of idle time to wait before starting the idle pattern")
+	viper.BindPFlag("typing.idle-period", typingPatternCmd.Flags().Lookup("idle-period"))
+
 	typingPatternCmd.Args = func(cmd *cobra.Command, _ []string) (err error) {
 		typingPattern.InputEventID = viper.GetString("typing.input-event-id")
+		typingPattern.CountAllKeys = viper.GetBool("typing.all-keys")
 		typingPattern.IdlePattern, err = getIdlePattern(cmd, viper.GetString("typing.idle"))
 		return
 	}
@@ -55,26 +70,17 @@ func addPatternCmd(short string, pattern patterns.Pattern) *cobra.Command {
 		Use:   pattern.GetBase().Name,
 		Short: short,
 		PreRunE: func(cmd *cobra.Command, _ []string) error {
-			err := checkAndSetPidPath(pidpath)
-			if err != nil {
+			if err := checkAndSetPidPath(pidpath); err != nil {
 				return fail(11, err)
 			}
-
-			err = beNice(priority)
-			if err != nil {
+			if err := beNice(priority); err != nil {
 				return fail(12, err)
 			}
-
-			plog := log.With().Str("cmd", cmd.Name()).Logger()
-			plog.Info().Msg("starting")
-
-			basePattern.Ctx = cmd.Context()
-			basePattern.Log = &plog
-			return nil
+			return startIPCServer(cmd.Context())
 		},
 		RunE: func(cmd *cobra.Command, _ []string) error {
 			basePattern.Delay = viper.GetDuration(basePattern.Name + ".delay")
-			return pattern.Run()
+			return pattern.Run(cmd.Context(), &log.Logger)
 		},
 		PostRun: func(_ *cobra.Command, _ []string) {
 			os.Remove(pidpath)
@@ -119,6 +125,11 @@ func checkAndSetPidPath(pidpath string) error {
 			return fmt.Errorf("unable to parse contents of %s: %w", pidpath, err)
 		}
 
+		if otherPid == os.Getpid() {
+			// just ourselves, probably running an IPC command
+			return nil
+		}
+
 		err = syscall.Kill(otherPid, 0)
 		if err == nil || err.(syscall.Errno) == syscall.EPERM {
 			// if EPERM, process is owned by another user, probably root
@@ -134,6 +145,7 @@ func checkAndSetPidPath(pidpath string) error {
 			return fmt.Errorf("unable to read %s: %w", pidpath, err)
 		}
 	}
+
 	err = os.WriteFile(pidpath, []byte(fmt.Sprint(os.Getpid())), 0666)
 	if err != nil {
 		return fmt.Errorf("unable to write to %s: %w", pidpath, err)
@@ -149,4 +161,15 @@ func beNice(priority int) error {
 		return fmt.Errorf("unable to set nice level %d: %w", priority, err)
 	}
 	return nil
+}
+
+func startIPCServer(ctx context.Context) error {
+	svr := ipc.IPCServer{
+		Ctx:  ctx,
+		Path: filepath.Join(os.TempDir(), buildinfo.Name+".sock"),
+		Cmd:  rootCmd,
+		Log:  &log.Logger,
+	}
+
+	return svr.Start()
 }
